feat:implementing cart feature with new api

+++ lib/features/domain/entities/product_entities/product_to_rfq_entity.dart
@@ -4,7 +4,7 @@ class ProductToRfq {
   final String productImage;
   final String hsCode;
   final String casNumber;
-  double? quantity;
+  int? quantity;
   String? unit;
 
   ProductToRfq(
